requirements.txt
torch
transformers
pytesseract
pillow
datasets
gtts
streamlit
playsound
